Refactored Uuidtext to stand-alone file and added tests
